initial file restructure
